// backend/src/generated/model/tool.ts
/**
 * Pooper API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { RegisterMcpToolRequestParametersInner } from './registerMcpToolRequestParametersInner';

export class Tool {
    'id': string;
    'name': string;
    'description'?: string;
    'serverId': string;
    'type'?: Tool.TypeEnum = Tool.TypeEnum.Custom;
    'parameters'?: Array<RegisterMcpToolRequestParametersInner>;
    'metadata'?: { [key: string]: any; };
    'createdAt'?: Date;
    'updatedAt'?: Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "serverId",
            "baseName": "serverId",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "Tool.TypeEnum"
        },
        {
            "name": "parameters",
            "baseName": "parameters",
            "type": "Array<RegisterMcpToolRequestParametersInner>"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "{ [key: string]: any; }"
        },
        {
            "name": "createdAt",
            "baseName": "createdAt",
            "type": "Date"
        },
        {
            "name": "updatedAt",
            "baseName": "updatedAt",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return Tool.attributeTypeMap;
    }
}

export namespace Tool {
    export enum TypeEnum {
        System = <any> 'system',
        Plugin = <any> 'plugin',
        Custom = <any> 'custom'
    }
}